+++ internal/pkg/oci/defaults.go
@@ -6,7 +6,7 @@ import (
 	warewulfconf "github.com/warewulf/warewulf/internal/pkg/config"
 )
 
-var defaultCachePath = filepath.Join(warewulfconf.Get().Warewulf.DataStore, "/container-cache/oci/")
+var defaultCachePath = filepath.Join(warewulfconf.Get().Paths.Datadir, "/container-cache/oci/")
 
 const (
 	blobPrefix   = "blobs"