+++ CHANGELOG.md
@@ -81,6 +81,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Provide detected kernel version to overlay templates. #1556
 - Bump github.com/containers/storage from 1.53.0 to 1.55.2 #1316, #892
 - Process nodes.conf path dynamically from config. #1595, #1596, #1569
+- Split overlays into distribution and site overlays. #831
 
 ### Removed